OSD and Force sensor update
Updated so that HX711 can display two force measurements (alternately read) 
Also added a an OSD for the read values, corrected some display issues so alternate measurements are now displayed alternately. This is important with the HX711 as the update rate with two senors is slow.
Default single channel sampling period is 500ms with this version.