Updated display routines
This commit has reinstated S-March's multiple ADC read before sending to the HTML system. I had broken this in earlier versions so only one adc reading got sent per HTML plot update. This is now fixed, and multiple ADC readings are now sent in batches to the HTML.

However, the ADC cannot be read whilst the HTML is being updated, so code has been added that shows a blank line in the scope plot where no samples have been made. 

when turned on for the first time, the scope screen can have strange display, but this sorts itself out once an adequate number of adc sample have been made and sent. 

The code supports two HX711 channels, but when reading two channels simultaneously, the sampling rate is VERY slow (to allow the HX711 to swap channels) 
Code in the HTML now sets the sample rate depending on what is being read: 
The internal ADC is read at 3ms intervals. 
If the HX711 is on, sample rate is reduced to 100ms. With both HX711 channels, sample rate is 1500ms, but code in